Optimization destroyed everything

+++ srcs/minecraft/Location.cpp
@@ -6,7 +6,7 @@
 /*   By: scraeyme <scraeyme@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/19 17:00:39 by scraeyme          #+#    #+#             */
-/*   Updated: 2025/04/22 00:41:04 by scraeyme         ###   ########.fr       */
+/*   Updated: 2025/04/22 20:13:29 by scraeyme         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ Location::Location()
 	this->_z = 0.0f;
 }
 
-Location::Location(float x, float y, float z)
+Location::Location(double x, double y, double z)
 {
 	this->_x = x;
 	this->_y = y;
@@ -31,10 +31,10 @@ bool Location::operator==(const Location &obj)
 	return (this->_x == obj._x && this->_y == obj._y && this->_z == obj._z);
 }
 
-float Location::getX() const { return this->_x; }
-float Location::getY() const { return this->_y; }
-float Location::getZ() const { return this->_z; }
+double Location::getX() const { return this->_x; }
+double Location::getY() const { return this->_y; }
+double Location::getZ() const { return this->_z; }
 
-void Location::setX(float x) { this->_x = x; }
-void Location::setY(float y) { this->_y = y; }
-void Location::setZ(float z) { this->_z = z; }
+void Location::setX(double x) { this->_x = x; }
+void Location::setY(double y) { this->_y = y; }
+void Location::setZ(double z) { this->_z = z; }
